Addressing more feedback.
